+++ lib/azure_upload_file.dart
@@ -5,6 +5,7 @@ import "dart:convert";
 
 import "package:async/async.dart";
 import "package:azure_upload_file/src/azure_storage.dart";
+import "package:azure_upload_file/src/dto/azure_progress_message.dart";
 import "package:cross_file/cross_file.dart";
 import "package:crypto/crypto.dart";
 import "package:flutter/foundation.dart";
@@ -13,30 +14,7 @@ import "package:rxdart/rxdart.dart";
 import "package:shared_preferences/shared_preferences.dart";
 
 export "src/azure_storage.dart";
-
-class AzureProgressMessage {
-  final double progress;
-  final Object? error;
-  final bool completed;
-
-  AzureProgressMessage({
-    required this.progress,
-    this.error,
-    this.completed = false,
-  });
-
-  @override
-  bool operator ==(covariant AzureProgressMessage other) {
-    if (identical(this, other)) return true;
-
-    return other.progress == progress &&
-        other.error == error &&
-        other.completed == completed;
-  }
-
-  @override
-  int get hashCode => progress.hashCode ^ error.hashCode ^ completed.hashCode;
-}
+export "src/dto/azure_progress_message.dart";
 
 class AzureUploadFile {
   late SharedPreferences _prefs;
@@ -57,24 +35,33 @@ class AzureUploadFile {
 
   AzureUploadFile();
 
+  // this is the upload stream returned when the startUpload is called
+  // This stream combines the progress stream of the AzureStorage
+  // with the error stream
   Stream<AzureProgressMessage> get _progressStream =>
       CombineLatestStream.combine2(
+        // when the progress stream emits, we store the upload status
+        // to correctly resume it in case of pause/error
         _azureStorage!.progressStream.doOnData((data) async {
           await _prefs.setString(_progressMapKey, jsonEncode(data));
         }),
         _errorSubj!.stream,
         (progressMap, e) {
+          // calculate the overall progress
           double progress = (progressMap as Map<String, int>).isEmpty
               ? 0
               : progressMap.values.reduce((a, b) => a + b) /
                   _azureStorage!.fileSize;
           progress = double.parse(progress.toStringAsFixed(2));
           if (e != null) debugPrint("error $e");
+          // emit the progress and the error if present
           return AzureProgressMessage(progress: progress, error: e);
         },
       ).distinct();
 
-  void emitError(Object? e) {
+  // all the errors generated by this library
+  // should be emitted with this method to correctly send them to the progress stream
+  void _emitError(Object? e) {
     _errorSubj!.add(e);
   }
 
@@ -95,7 +82,7 @@ class AzureUploadFile {
       );
       await _prefs.setString(_sasLinkKey, sasLink);
     } catch (e) {
-      emitError(e);
+      _emitError(e);
     }
   }
 
@@ -124,7 +111,7 @@ class AzureUploadFile {
         resume: resume,
       );
     } catch (e) {
-      emitError(e);
+      _emitError(e);
     }
 
     return _progressStream;
@@ -157,7 +144,7 @@ class AzureUploadFile {
         resume: true,
       );
     } catch (e) {
-      emitError(e);
+      _emitError(e);
     }
   }
 
@@ -244,18 +231,25 @@ class AzureUploadFile {
 
         nextBytes = await fileReader.readBytes(_chunkSize);
       }
+      // if nextBytes is empty, then we completed the upload
+      // we need to close the progress streams and prepare
+      // for next upload
       if (nextBytes.isEmpty == true) {
         _deletePrefs();
+        // close file
         await fileReader.cancel();
+        // close azure storage stream
         _azureStorage!.closeStream();
+        // close AND DELETE the error stream (this permit the use of the library inside an isolate)
         _errorSubj?.close();
         _errorSubj = null;
+        // reset isPaused value
         _isPaused = true;
         debugPrint("AzureUploadFile: completed");
       }
       debugPrint("AzureUploadFile: exited");
     } catch (e) {
-      emitError(e);
+      _emitError(e);
     }
   }
 }
